Extract MCP OAuth surface into routes/auth/mcp_oauth.py
Move the 8 OAuth/OIDC routes (.well-known/openid-configuration,
.well-known/oauth-authorization-server, jwks.json, register, consent,
browser-session, authorize, token) plus their helpers from
routes/auth/__init__.py into a new mcp_oauth.py module exposing a
register_mcp_oauth_routes(auth_blp, *, deps, ...) function.

The two DB-coupled bridge helpers
(_ensure_linked_zitadel_subject_for_user, _linked_zitadel_subject_for_email)
stay in __init__.py and are injected into the new registration function as
callbacks. Everything else moves: PKCE/scope/redirect-URI validation,
refresh-token rotation, signing-key rollover, stale-client sweep,
authorization-code minting, the OAuth user-grant model wiring.

Behavior identical: every closure body is a verbatim lift, route paths
unchanged, abort codes/descriptions unchanged. Pre-existing dead helper
_oauth_response_with_pending_cookie_cleared (defined but never called)
is dropped — verified by grep across the backend.

Module constants _OAUTH_LOOPBACK_REDIRECT_HOSTS, _DCR_UNUSED_GRACE_HOURS,
_DCR_INACTIVE_DAYS moved alongside the code that uses them. Cleaned up
14 imports from __init__.py that were OAuth-only (escape,
urlsafe_b64encode, urlparse/urlunparse, and_/or_/update,
all mcp_oauth_runtime helpers, oauth cookie helpers, mcp_resource_url,
mcp_supported_scopes).
